Extract the feedback and improve routes out of submissions.ts
`POST /api/submissions/:token/feedback` and `.../improve` are the same act at
two points in a game's life: a creator asking for a change. Feedback steers a
draft that is still being built; improve opens a new round on a game that has
already shipped, because publishing is terminal. They are neighbours, not one
routine, so they land as two files rather than one — and at 276 and 203 lines
of route body, one file would also have cleared the 500-line cap.

  creation/feedback-routes.ts    registerFeedbackRoutes    (14 deps)
  creation/improve-routes.ts     registerImproveRoutes     (14 deps)

Both route bodies are unchanged. Proved by diffing each against master with
comments and blank lines stripped: byte-identical, 200 and 174 code lines.

`FeedbackRequestSchema` and `REFERENCE_IMAGES_BODY_LIMIT_BYTES` were private to
submissions.ts and are read by both new files and by the create route that
stays. They move to `creation/feedback-request.ts` rather than being exported
from submissions.ts, which the new files import — exporting them would have
been a runtime import cycle. `ReferenceImagesSchema` is named there too, so the
create route stops reaching for a bare `z.array(referenceImageSchema)`.

The `ResumeOutcome` and improvement-round return types are declared
structurally in the new files, for the same cycle reason as #1048.

submissions.ts: 3171 -> 2665 lines, ceiling resealed to match. Since #1046 the
file has lost 858 lines and five route clusters.

The moderation-metrics guard test listed every module that calls
`logModerationRejection`; two call sites moved, so it now names the new files.
Its own line ceiling is raised by two for those entries — a pinned inventory
grows when the modules it inventories split, and there is nothing to shrink.
