+++ eslint-rules/module-size-baseline.json
@@ -374,7 +374,7 @@
     "apps/api/src/platform/knowledge-metrics.test.ts": 60,
     "apps/api/src/platform/knowledge-metrics.ts": 23,
     "apps/api/src/platform/localize-intake.ts": 95,
-    "apps/api/src/platform/moderation-metrics.test.ts": 180,
+    "apps/api/src/platform/moderation-metrics.test.ts": 182,
     "apps/api/src/platform/moderation-metrics.ts": 86,
     "apps/api/src/platform/moderation-terms.ts": 120,
     "apps/api/src/platform/moderation.test.ts": 228,
@@ -437,7 +437,7 @@
     "apps/api/src/static-serving.test.ts": 95,
     "apps/api/src/store-firestore.test.ts": 710,
     "apps/api/src/submissions.test.ts": 7368,
-    "apps/api/src/submissions.ts": 3171,
+    "apps/api/src/submissions.ts": 2665,
     "apps/api/src/tab-complete-deploy-flag.test.ts": 26,
     "apps/api/src/telemetry-trends.route.test.ts": 75,
     "apps/api/src/telemetry/chat-agent-metrics.test.ts": 53,
